openclaw-skills/gtv-assessment/scripts/requirements.txt
# GTV Assessment Skill 脚本依赖
# generate_report.py / compare_paths.py — 仅使用标准库
# generate_pdf_report.py — 需要 reportlab
# parse_resume.py — 需要 pdfminer.six, python-docx
reportlab>=4.0
pdfminer.six>=20221105
python-docx>=1.0
